net: ipv6: allow choosing optimistic addresses with use_optimistic
The use_optimistic sysctl makes optimistic IPv6 addresses equivalent to
preferred addresses for source address selection (e.g., when calling
connect()), but it does not allow an application to bind to optimistic
addresses. This behaviour is inconsistent - for example, it doesn't make
sense for bind() to an optimistic address fail with EADDRNOTAVAIL, but
connect() to choose that address outgoing address on the same socket.

Bug: 17769720
Bug: 18609055

Change-Id: I9de0d6c92ac45e29d28e318ac626c71806666f13
Signed-off-by: Erik Kline <ek@google.com>
Signed-off-by: Lorenzo Colitti <lorenzo@google.com>